Add programme pdf and link from homepage
The programme is a pdf document, exported from Word. Both the pdf and Word documents are in content/

The `build.py` will do the usual pandoc-ing of content/home.md to docs/index.html and now in addition, it will cp the programme.pdf from content/ to docs/

The programme.pdf could have been manually put in docs/ in the first place, but I prefer to do it this way: the source documents are in content/ and the files to be served are in docs/, which are always regenerated with `build.py`.

Note `build.py` now uses shutil from the cp-ing, but this will work cross platform I believe.